requirements.txt
torch~=1.13.1+cu117
numpy~=1.26.2
pyyaml~=6.0.1
tqdm~=4.66.1
pillow~=10.1.0
torchvision~=0.14.1+cu117